fit/backend/requirements.txt
fastapi
uvicorn
google-auth-oauthlib
google-api-python-client
python-dotenv
requests
pandas
numpy
